Updated.
For MVP, we still need a shop to work in order to have "Has a currency system which utilizes arithmetic operations".  Once we have that, we have MVP.  I have completed one stretch goal, reading save files.  If we have a working shop, we would have an 80%.  If we had a working shop and randomized dungeon, we would have a 90%, if we had NPCs, we could have 100%.  I will make a shop that will edit your gold, so that we at least have MVP.  If you guys could work on the dungeon and NPCs.