Working on saving progresss
Added a user class and a moderator class that both extend to abstractUser interface, which will have username and shared things in it.  Working on saving files based on team name, added team name to user, it displays when you view your team.  About halfway done with saving and loading files work.